[Refactor] Change the component folder location
